+++ oauth2_provider/backends.py
@@ -1,14 +1,17 @@
 """Backward-compatible import shim.
 
-``oauth2_provider.backends`` moved to ``oauth2_provider.resource_server.backends`` when
-the package was reorganized by OAuth2 role. Importing from this old path still
-works but is deprecated and will be removed in django-oauth-toolkit 4.0.
+``oauth2_provider.backends`` has moved to
+``oauth2_provider.resource_server.backends``.
+
+Importing from this old path still works but is deprecated and will be removed
+in django-oauth-toolkit 4.0.
 """
 
 import sys
 import warnings
 
 from oauth2_provider.resource_server import backends as _moved
+
 
 warnings.warn(
     "oauth2_provider.backends has moved to oauth2_provider.resource_server.backends. "
@@ -18,4 +21,6 @@ warnings.warn(
     stacklevel=2,
 )
 
+# Alias the old module name to the moved module so every attribute (including
+# private, underscore-prefixed names) resolves with identical object identity.
 sys.modules[__name__] = _moved
