+++ oauth2_provider/cimd.py
@@ -1,8 +1,10 @@
 """Backward-compatible import shim.
 
-``oauth2_provider.cimd`` moved to ``oauth2_provider.authorization_server.cimd`` when the package was reorganized
-by OAuth2 role. Importing from this old path still works but is deprecated and
-will be removed in django-oauth-toolkit 4.0.
+``oauth2_provider.cimd`` has moved to
+``oauth2_provider.authorization_server.cimd``.
+
+Importing from this old path still works but is deprecated and will be removed
+in django-oauth-toolkit 4.0.
 """
 
 import sys
@@ -10,9 +12,11 @@ import warnings
 
 from oauth2_provider.authorization_server import cimd as _moved
 
+
 warnings.warn(
-    "oauth2_provider.cimd has moved to oauth2_provider.authorization_server.cimd. The old import path is "
-    "deprecated and will be removed in django-oauth-toolkit 4.0.",
+    "oauth2_provider.cimd has moved to oauth2_provider.authorization_server.cimd. "
+    "The old import path is deprecated and will be removed in "
+    "django-oauth-toolkit 4.0.",
     DeprecationWarning,
     stacklevel=2,
 )
