+++ oauth2_provider/decorators.py
@@ -1,14 +1,17 @@
 """Backward-compatible import shim.
 
-``oauth2_provider.decorators`` moved to ``oauth2_provider.resource_server.decorators`` when
-the package was reorganized by OAuth2 role. Importing from this old path still
-works but is deprecated and will be removed in django-oauth-toolkit 4.0.
+``oauth2_provider.decorators`` has moved to
+``oauth2_provider.resource_server.decorators``.
+
+Importing from this old path still works but is deprecated and will be removed
+in django-oauth-toolkit 4.0.
 """
 
 import sys
 import warnings
 
 from oauth2_provider.resource_server import decorators as _moved
+
 
 warnings.warn(
     "oauth2_provider.decorators has moved to oauth2_provider.resource_server.decorators. "
@@ -18,4 +21,6 @@ warnings.warn(
     stacklevel=2,
 )
 
+# Alias the old module name to the moved module so every attribute (including
+# private, underscore-prefixed names) resolves with identical object identity.
 sys.modules[__name__] = _moved
