+++ oauth2_provider/oauth2_backends.py
@@ -1,8 +1,10 @@
 """Backward-compatible import shim.
 
-``oauth2_provider.oauth2_backends`` moved to ``oauth2_provider.core.backends_oauthlib`` when the package was reorganized
-by OAuth2 role. Importing from this old path still works but is deprecated and
-will be removed in django-oauth-toolkit 4.0.
+``oauth2_provider.oauth2_backends`` has moved to
+``oauth2_provider.core.backends_oauthlib``.
+
+Importing from this old path still works but is deprecated and will be removed
+in django-oauth-toolkit 4.0.
 """
 
 import sys
@@ -10,9 +12,11 @@ import warnings
 
 from oauth2_provider.core import backends_oauthlib as _moved
 
+
 warnings.warn(
-    "oauth2_provider.oauth2_backends has moved to oauth2_provider.core.backends_oauthlib. The old import path is "
-    "deprecated and will be removed in django-oauth-toolkit 4.0.",
+    "oauth2_provider.oauth2_backends has moved to oauth2_provider.core.backends_oauthlib. "
+    "The old import path is deprecated and will be removed in "
+    "django-oauth-toolkit 4.0.",
     DeprecationWarning,
     stacklevel=2,
 )
