+++ oauth2_provider/oauth2_validators.py
@@ -1,19 +1,15 @@
 import base64
 import binascii
 import hashlib
-import http.client
 import inspect
 import json
 import logging
-import posixpath
-import urllib.parse
 import uuid
 from collections import OrderedDict
-from datetime import datetime, timedelta
+from datetime import timedelta
 from datetime import timezone as datetime_timezone
 from urllib.parse import unquote_plus
 
-import requests
 from django.conf import settings
 from django.contrib.auth import authenticate, get_user_model
 from django.contrib.auth.hashers import check_password, identify_hasher
@@ -34,6 +30,7 @@ from oauthlib.openid import RequestValidator
 from .authorization_server import cimd
 from .authorization_server.bcp import bcp_compliant
 from .core.exceptions import FatalClientError
+from .core.scopes import get_scopes_backend
 from .models import (
     AbstractApplication,
     get_access_token_model,
@@ -44,114 +41,13 @@ from .models import (
     refresh_token_expire_timedelta,
     revoke_access_token,
 )
-from .core.scopes import get_scopes_backend
+from .resource_server.validators import (  # noqa: F401  (re-exported for backward compatibility)
+    ResourceServerValidatorMixin,
+    _parse_and_validate_uri,
+    is_valid_resource_uri,
+    validate_resource_as_url_prefix,
+)
 from .settings import oauth2_settings
-from .core.utils import get_timezone
-
-
-# Default ports used to normalize URIs before comparison, so that e.g.
-# "https://api.example.com:443/foo" and "https://api.example.com/foo" match.
-_SCHEME_DEFAULT_PORTS = {"http": 80, "https": 443}
-
-
-def _parse_and_validate_uri(uri):
-    """Parse a URI and return (scheme, hostname, port, path) or None if invalid.
-
-    Paths are normalized to resolve dot segments (RFC 3986 Section 5.2.4).
-    For http and https URIs, an omitted port is normalized to the scheme default
-    (RFC 3986 Section 6.2.3); for other schemes it remains None.
-    URIs with userinfo or fragment components are rejected. A query component is
-    accepted (RFC 8707 allows one on resource indicators) but is not part of the
-    returned tuple, so it plays no role in matching.
-
-    A non-string ``uri`` (e.g. a non-string element in a JSON ``resource`` array)
-    returns None rather than raising, so callers fail closed with a clean
-    ``invalid_target`` error instead of a TypeError / 500.
-    """
-    if not isinstance(uri, str):
-        return None
-    parsed = urllib.parse.urlsplit(uri)
-    # "is not None" catches empty-but-present components ("https://@example.com",
-    # "https://example.com/#") that truthiness checks would let through.
-    if parsed.username is not None or parsed.password is not None:
-        return None
-    if parsed.fragment or "#" in uri:
-        return None
-    if not parsed.scheme or not parsed.hostname:
-        return None
-    scheme = parsed.scheme.lower()
-    try:
-        port = parsed.port
-    except ValueError:
-        # Non-numeric or out-of-range port
-        return None
-    if port is None:
-        port = _SCHEME_DEFAULT_PORTS.get(scheme)
-    path = posixpath.normpath(parsed.path or "/")
-    return (scheme, parsed.hostname.lower(), port, path)
-
-
-def is_valid_resource_uri(uri):
-    """Return True if ``uri`` is acceptable as an RFC 8707 resource indicator.
-
-    Accepted values are absolute URIs with a scheme and host, without userinfo
-    or fragment components (e.g. ``https://api.example.com/path``).
-    """
-    return _parse_and_validate_uri(uri) is not None
-
-
-def validate_resource_as_url_prefix(request_uri, audiences):
-    """
-    Default resource validator using URL prefix matching (RFC 8707).
-
-    Validates that the request URI matches one of the token's audience claims
-    using prefix matching. The audience URI acts as a base URI that the request
-    must start with.
-
-    URIs are parsed and compared component-by-component to prevent bypasses
-    via userinfo injection or authority confusion.
-
-    Examples:
-        - Token audience: "https://api.example.com/foo"
-        - Matches: "https://api.example.com/foo"
-        - Matches: "https://api.example.com/foo/"
-        - Matches: "https://api.example.com/foo/bar"
-        - Rejects: "https://other.example.com/foo/bar"
-        - Rejects: "https://api.example.com/bar"
-        - Rejects: "https://api.example.com/food-blog"
-        - Rejects: "https://api.example.com@evil.com/foo"
-
-    Both ``request_uri`` and the audience values must be absolute URIs with a
-    scheme and host; other absolute-URI forms (e.g. URNs) never match.
-
-    :param request_uri: String URI of the current request (without query string)
-    :param audiences: List of audience URI strings from token
-    :return: True if token is valid for this request, False otherwise
-    """
-    if not audiences:
-        return True
-
-    request_parts = _parse_and_validate_uri(request_uri)
-    if request_parts is None:
-        return False
-
-    req_scheme, req_host, req_port, req_path = request_parts
-    req_path_normalized = req_path.rstrip("/") + "/"
-
-    for audience in audiences:
-        aud_parts = _parse_and_validate_uri(audience)
-        if aud_parts is None:
-            continue
-
-        aud_scheme, aud_host, aud_port, aud_path = aud_parts
-        if (req_scheme, req_host, req_port) != (aud_scheme, aud_host, aud_port):
-            continue
-
-        aud_path_normalized = aud_path.rstrip("/") + "/"
-        if req_path_normalized.startswith(aud_path_normalized):
-            return True
-
-    return False
 
 
 log = logging.getLogger("oauth2_provider")
@@ -181,7 +77,7 @@ RefreshToken = get_refresh_token_model()
 UserModel = get_user_model()
 
 
-class OAuth2Validator(RequestValidator):
+class OAuth2Validator(ResourceServerValidatorMixin, RequestValidator):
     # Return the given claim only if the given scope is present.
     # Extended as needed for non-standard OIDC claims/scopes.
     # Override by setting to None to ignore scopes.
@@ -512,184 +408,6 @@ class OAuth2Validator(RequestValidator):
         """
         user, _ = UserModel.objects.get_or_create(**{UserModel.USERNAME_FIELD: content["username"]})
         return user
-
-    def _get_token_from_authentication_server(
-        self, token, introspection_url, introspection_token, introspection_credentials
-    ):
-        """Use external introspection endpoint to "crack open" the token.
-        :param introspection_url: introspection endpoint URL
-        :param introspection_token: Bearer token
-        :param introspection_credentials: Basic Auth credentials (id,secret)
-        :return: :class:`models.AccessToken`
-
-        Some RFC 7662 implementations (including this one) use a Bearer token while others use Basic
-        Auth. Depending on the external AS's implementation, provide either the introspection_token
-        or the introspection_credentials.
-
-        If the resulting access_token identifies a username (e.g. Authorization Code grant), add
-        that user to the UserModel. Also cache the access_token up until its expiry time or a
-        configured maximum time.
-
-        """
-        headers = None
-        if introspection_token:
-            headers = {"Authorization": "Bearer {}".format(introspection_token)}
-        elif introspection_credentials:
-            client_id = introspection_credentials[0].encode("utf-8")
-            client_secret = introspection_credentials[1].encode("utf-8")
-            basic_auth = base64.b64encode(client_id + b":" + client_secret)
-            headers = {"Authorization": "Basic {}".format(basic_auth.decode("utf-8"))}
-
-        try:
-            response = requests.post(introspection_url, data={"token": token}, headers=headers)
-        except requests.exceptions.RequestException:
-            log.exception("Introspection: Failed POST to %r in token lookup", introspection_url)
-            return None
-
-        # Log an exception when response from auth server is not successful
-        if response.status_code != http.client.OK:
-            log.exception(
-                "Introspection: Failed to get a valid response "
-                "from authentication server. Status code: {}, "
-                "Reason: {}.".format(response.status_code, response.reason)
-            )
-            return None
-
-        try:
-            content = response.json()
-        except ValueError:
-            log.exception("Introspection: Failed to parse response as json")
-            return None
-
-        if "active" in content and content["active"] is True:
-            if "username" in content:
-                user = self.get_or_create_user_from_content(content)
-            else:
-                user = None
-
-            max_caching_time = datetime.now(tz=datetime_timezone.utc) + timedelta(
-                seconds=oauth2_settings.RESOURCE_SERVER_TOKEN_CACHING_SECONDS
-            )
-
-            if "exp" in content:
-                expires = datetime.fromtimestamp(content["exp"], tz=datetime_timezone.utc)
-                exp_time_zone = oauth2_settings.AUTHENTICATION_SERVER_EXP_TIME_ZONE
-                if exp_time_zone != "UTC":
-                    # Deprecated AUTHENTICATION_SERVER_EXP_TIME_ZONE workaround: reinterpret the
-                    # exp wall-clock time as being in the configured (non-UTC) time zone.
-                    expires = make_aware(expires.replace(tzinfo=None), timezone=get_timezone(exp_time_zone))
-                if expires > max_caching_time:
-                    expires = max_caching_time
-            else:
-                expires = max_caching_time
-
-            scope = content.get("scope", "")
-
-            if not settings.USE_TZ:
-                expires = timezone.make_naive(expires, expires.tzinfo)
-
-            # RFC 8707: Map introspection 'aud' claim to resource field.
-            # RFC 7662 defines 'aud' as a string or array of strings. 'aud' is a
-            # security restriction, so a malformed value must fail closed: treating
-            # it as unrestricted would let the token through on any resource.
-            aud = content.get("aud", [])
-            if isinstance(aud, str):
-                aud = [aud]
-            elif not isinstance(aud, list) or not all(isinstance(entry, str) for entry in aud):
-                log.warning("Rejecting token: malformed 'aud' claim in introspection response: %r", aud)
-                return None
-
-            token_checksum = hashlib.sha256(token.encode("utf-8")).hexdigest()
-            # Respect hashed-at-rest storage (RFC 9700): the resource-server token cache
-            # is looked up by checksum, so it must not persist the cleartext token when
-            # COMPLIANT_BCP_RFC9700_TOKEN_STORAGE is enabled.
-            stored_token = "" if oauth2_settings.COMPLIANT_BCP_RFC9700_TOKEN_STORAGE else token
-            access_token, _created = AccessToken.objects.update_or_create(
-                token_checksum=token_checksum,
-                defaults={
-                    "token": stored_token,
-                    "user": user,
-                    "application": None,
-                    "scope": scope,
-                    "expires": expires,
-                    "resource": aud,
-                },
-            )
-
-            return access_token
-
-    def validate_bearer_token(self, token, scopes, request):
-        """
-        When users try to access resources, check that provided token is valid
-        """
-        if not token:
-            return False
-
-        introspection_url = oauth2_settings.RESOURCE_SERVER_INTROSPECTION_URL
-        introspection_token = oauth2_settings.RESOURCE_SERVER_AUTH_TOKEN
-        introspection_credentials = oauth2_settings.RESOURCE_SERVER_INTROSPECTION_CREDENTIALS
-
-        access_token = self._load_access_token(token)
-
-        # if there is no token or it's invalid then introspect the token if there's an external OAuth server
-        if not access_token or not access_token.is_valid(scopes):
-            if introspection_url and (introspection_token or introspection_credentials):
-                access_token = self._get_token_from_authentication_server(
-                    token, introspection_url, introspection_token, introspection_credentials
-                )
-
-        if access_token and access_token.is_valid(scopes):
-            # RFC 8707: Only resource-restricted tokens are audience-checked, so
-            # unrestricted tokens keep working with minimal request objects that
-            # carry no uri. Restricted tokens fail closed when the uri is absent.
-            if access_token.resource:
-                # request.uri is the full URI from the oauthlib Request object
-                request_uri = (getattr(request, "uri", None) or "").split("?")[0]
-                if not access_token.allows_audience(request_uri):
-                    request.oauth2_error = OrderedDict(
-                        [
-                            ("error", "invalid_token"),
-                            (
-                                "error_description",
-                                _("The access token is not valid for this resource."),
-                            ),
-                        ]
-                    )
-                    return False
-
-            # Reject tokens whose application is no longer usable, mirroring the
-            # is_usable() check the issuance path performs in _load_application.
-            # The default is_usable() returns True, so this is a no-op unless a
-            # swapped Application model overrides it (e.g. to disable an app or
-            # gate on request attributes). See #1260.
-            application = access_token.application
-            if application is not None and not application.is_usable(request):
-                request.oauth2_error = OrderedDict(
-                    [
-                        ("error", "invalid_token"),
-                        ("error_description", _("The access token is invalid.")),
-                    ]
-                )
-                return False
-
-            request.client = access_token.application
-            request.user = access_token.user
-            request.scopes = list(access_token.scopes)
-
-            # this is needed by django rest framework
-            request.access_token = access_token
-            return True
-        else:
-            self._set_oauth2_error_on_request(request, access_token, scopes)
-            return False
-
-    def _load_access_token(self, token):
-        token_checksum = hashlib.sha256(token.encode("utf-8")).hexdigest()
-        return (
-            AccessToken.objects.select_related("application", "user")
-            .filter(token_checksum=token_checksum)
-            .first()
-        )
 
     def validate_code(self, client_id, code, client, request, *args, **kwargs):
         try:
