+++ oauth2_provider/resource_server/middleware.py
@@ -4,8 +4,8 @@ import logging
 from django.contrib.auth import authenticate
 from django.utils.cache import patch_vary_headers
 
-from oauth2_provider.models import get_access_token_model
 from oauth2_provider.core.utils import parse_bearer_token
+from oauth2_provider.models import get_access_token_model
 
 
 log = logging.getLogger(__name__)
